yocto-bsp: drop MIPS Edgerouter support
As per ongoing discussions, there is a general desire to try and update
our hardware reference platforms when an appropriate (newer) substitute
is available.  However, in the case of MIPS, there isn't an obvious one
to choose.  But continuing to maintain support for an old board that
hasn't been available for purchase for years doesn't make sense either.

So we are dropping support for the Edgerouter for future releases.  The
MIPS architecture will still get coverage via QEMU build/boot testing.

Signed-off-by: Paul Gortmaker <paulg@kernel.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>