base-files: add usage warning to motd
The Poky distro is explicitly a _reference_ distribution for _testing_
and _development_ purposes.  It enables most hardware and software
features so that they can be tested, but this also means that
from a security point of view the attack surface is very large.

We encourage anyone using OpenEmbedded for production use to create their
own distribution and not use Poky.  To encourage this behaviour further,
add a warning to /etc/motd when Poky is used so the developer will see it
when they log in.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>