genericarm64: depend on u-boot in testimage
genericarm64 supports runqemu, so testimage should work. However, out of
the box it does not:

runqemu - ERROR - BIOS .../tmp/deploy/images/genericarm64/u-boot.bin not found

That is because we make the user explicitly build the qemu-targetting u-boot
to avoid the confusing situation where they have real hardware but the
deploy directory contains a u-boot for qemu.

In automated testing situations, however, we can be a bit more helpful
and make testimage depend on u-boot. This will lead to u-boot binaries
being in deploy, but at this point the user is already running the images
inside a qemu.

Reported-by: James McGregor <James.McGregor2@arm.com>
Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>