local.conf.sample: use explicit IMAGE_FEATURES instead of debug-tweaks
debug-tweaks is vague and doesn't give any indication that the root user
can login over SSH without a password.  This behaviour is incredibly
dangerous if used unwittingly, so discourage it by using the underlying
features explicitly to spell out exactly what is being done.

This is not the complete set of features that debug-tweaks enables,
but I don't think we need to enable serial-autologin-root or
post-install-logging by default.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>