genericarm64.conf: allow overriding u-boot and qemuboot variables
Assignment in machine config can not be overwritten in local.conf
because former is evaluated later than latter. Use ?= conditional
assignment for u-boot and qemuboot variables so that they can be
overwritten in local.conf if a different test setup is wanted
for kernel, initrd and rootfs built for genericarm64 machine
config.

Signed-off-by: Mikko Rapeli <mikko.rapeli@linaro.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>